@@ -56,37 +56,38 @@
   },
   "private": true,
   "devDependencies": {
-    "@nrwl/cli": "13.4.1",
-    "@nrwl/eslint-plugin-nx": "13.4.1",
-    "@nrwl/jest": "13.4.1",
-    "@nrwl/js": "13.4.1",
-    "@nrwl/linter": "13.4.1",
-    "@nrwl/node": "13.4.1",
-    "@nrwl/tao": "13.4.1",
-    "@nrwl/workspace": "13.4.1",
-    "@types/jest": "27.0.2",
-    "@types/node": "14.14.33",
+    "@ngneat/transloco-keys-manager": "3.4.2",
+    "@ngneat/transloco-scoped-libs": "3.0.4",
+    "@nrwl/cli": "14.6.5",
+    "@nrwl/eslint-plugin-nx": "14.6.5",
+    "@nrwl/jest": "14.6.5",
+    "@nrwl/js": "14.6.5",
+    "@nrwl/linter": "14.6.5",
+    "@nrwl/node": "14.6.5",
+    "@nrwl/workspace": "14.6.5",
+    "@types/jest": "28.1.8",
+    "@types/node": "18.7.1",
     "@types/pg": "8.6.5",
-    "@types/recursive-readdir": "2.2.0",
+    "@types/recursive-readdir": "2.2.1",
     "@types/replace-ext": "2.0.0",
-    "@typescript-eslint/eslint-plugin": "5.3.0",
-    "@typescript-eslint/parser": "5.3.0",
+    "@typescript-eslint/eslint-plugin": "5.33.1",
+    "@typescript-eslint/parser": "5.33.1",
     "cp-cli": "2.0.0",
-    "eslint": "8.2.0",
+    "eslint": "8.15.0",
     "eslint-config-prettier": "8.1.0",
-    "jest": "27.2.3",
-    "log4js": "6.3.0",
-    "nestjs-console": "7.0.1",
-    "prettier": "2.3.1",
+    "jest": "28.1.3",
+    "log4js": "6.6.1",
+    "nestjs-console": "8.0.0",
+    "nx": "14.6.5",
+    "prettier": "2.7.1",
     "recursive-readdir": "2.2.2",
     "replace-ext": "2.0.0",
     "sort-paths": "1.1.1",
-    "standard-version": "9.3.2",
-    "testcontainers": "8.4.0",
-    "ts-jest": "27.0.5",
-    "typescript": "4.3.5",
-    "@ngneat/transloco-keys-manager": "3.4.1",
-    "@ngneat/transloco-scoped-libs": "3.0.4"
+    "standard-version": "9.5.0",
+    "testcontainers": "8.13.1",
+    "ts-jest": "28.0.8",
+    "ts-node": "10.9.1",
+    "typescript": "4.7.4"
   },
   "dependencies": {
     "@nestjs/common": "8.2.4",
@@ -97,8 +98,9 @@
     "gettext-extractor": "3.5.4",
     "i18next-conv": "12.1.1",
     "lodash.mergewith": "4.6.2",
-    "pg": "8.7.3",
-    "pg-promise": "10.11.1",
+    "pg": "8.8.0",
+    "pg-promise": "10.12.0",
+    "reflect-metadata": "^0.1.13",
     "source-map-support": "0.5.21",
     "tslib": "2.0.0"
   },
