@@ -39,7 +39,9 @@ export class Extracti18nService {
       this.logger.info('Process applications...');
       Object.keys(projects)
         .filter(
-          (projectName) => projects[projectName].projectType === 'application'
+          (projectName) =>
+            projects[projectName].projectType === 'application' ||
+            ( projects[projectName].sourceRoot?.substring(0,5) === 'apps/')
         )
         .forEach((projectName) => {
           this.logger.debug(projectName, projects[projectName].sourceRoot);
@@ -53,7 +55,9 @@ export class Extracti18nService {
       this.logger.info('Process libraries...');
       Object.keys(projects)
         .filter(
-          (projectName) => projects[projectName].projectType === 'library'
+          (projectName) =>
+            projects[projectName].projectType === 'library' ||
+            ( projects[projectName].sourceRoot?.substring(0,5) === 'libs/')
         )
         .forEach((projectName) => {
           this.logger.debug(projectName, projects[projectName].sourceRoot);
@@ -101,7 +105,8 @@ export class Extracti18nService {
     const scopedLibs = Object.keys(projects)
       .filter(
         (projectName) =>
-          projects[projectName].projectType === 'library' &&
+          (projects[projectName].projectType === 'library' ||
+            ( projects[projectName].sourceRoot?.substring(0,5) === 'libs/')) &&
           !projectName.includes('server')
       )
       .map((projectName) => projects[projectName].root);
@@ -109,7 +114,8 @@ export class Extracti18nService {
     const applications = Object.keys(projects)
       .filter(
         (projectName) =>
-          projects[projectName].projectType === 'application' &&
+          (projects[projectName].projectType === 'application' ||
+            ( projects[projectName].sourceRoot?.substring(0,5) === 'apps/')) &&
           !projectName.includes('server') &&
           !projectName.includes('-ms') &&
           !projectName.includes('e2e')
@@ -153,7 +159,8 @@ export class Extracti18nService {
     const scopedLibs = Object.keys(projects)
       .filter(
         (projectName) =>
-          projects[projectName].projectType === 'library' &&
+          (projects[projectName].projectType === 'library' ||
+            ( projects[projectName].sourceRoot?.substring(0,5) === 'libs/')) &&
           !projectName.includes('client')
       )
       .map((projectName) => projects[projectName].root);
@@ -161,7 +168,8 @@ export class Extracti18nService {
     const applications = Object.keys(projects)
       .filter(
         (projectName) =>
-          projects[projectName].projectType === 'application' &&
+          (projects[projectName].projectType === 'application' ||
+            ( projects[projectName].sourceRoot?.substring(0,5) === 'apps/')) &&
           !projectName.includes('client') &&
           !projectName.includes('e2e')
       )
